extract import related code to a hook
