+++ webapp/i18n/en.json
@@ -139,7 +139,6 @@
   "DSVJjB": "Currently running the <strong>{playbookTitle}</strong> playbook",
   "DXACD6": "Publish retrospective report and access the timeline",
   "DaHpK1": "Search for a channel",
-  "Dbex52": "The playbook import has failed. Please check that JSON is valid and try again.",
   "DnBhRg": "Add People",
   "DtCplA": "{numParticipants, plural, =1 {<b>#</b> participant} other {<b>#</b> participants}}",
   "E0LnBo": "You can select an option or specify a custom duration (\"2 weeks\", \"3 days 12 hours\", \"45 minutes\", ...)",
@@ -286,6 +285,7 @@
   "SVwJTM": "Export",
   "SXJ98n": "You will not be able to edit the retrospective report after publishing it. Do you want to publish the retrospective report?",
   "SmAUf9": "A reminder will be sent <b>{timestamp}</b>",
+  "Suyx6A": "The playbook import has failed. Please check that JSON is valid and try again.",
   "Sx3lHL": "Integer",
   "T5rX+W": "How often should an update be posted?",
   "TBez4r": "There are no playbooks to view. You don't have permission to create playbooks in this workspace.",
